Create 2. Data cleaning
Clean the dataset.

Create a list named chars_to_remove that contains the following characters: + , and $.
Create a list named cols_to_clean that contains the following column names: Installs and Price.
For each column in cols_to_clean in the apps DataFrame, replace each character in chars_to_remove with the empty string ''.
Note: Make sure to use an empty string '' and not a space character ' '
Finally, print a summary of the apps dataframe using the info() function. Observe the output.
Note: Notice that Installs and Price are still of type object and not int or float as we would have expected after removal of the special characters. We will solve this issue in the next task.