Create 1. Google Play Store apps and reviews
Import the data, drop duplicate rows, and inspect the data.

Load datasets/apps.csv into a DataFrame and assign it to the variable apps_with_duplicates.
Drop all duplicate rows from apps_with_duplicates and assign the result to apps.
Print the total number of apps.
Finally, display a random sample of 5 rows from apps.
Good to know
This project lets you apply the skills from Joining Data with pandas. We recommend that you take this course before starting this project.

The data for this project was scraped from the Google Play website. While there are many popular datasets for Apple App Store, there aren't many for Google Play apps, which is partially due to the increased difficulty in scraping the latter as compared to the former.